News_Agents/podcasts/requirements.txt
feedparser
requests
anthropic
openai-whisper
openai
